refactor: extrapolated the ticketview file into multiple components.
