+++ frontend/src/utils/initialsLogic.ts
@@ -1,0 +1,7 @@
+export const getInitials = (name: string) =>
+	name
+		.split(" ")
+		.map((n) => n[0])
+		.join("")
+		.toUpperCase()
+		.slice(0, 2);
